Moved BroadcastReceiver, simplified its actions
